Move tests only to where they are actually run
